Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.java
@@ -1,19 +1,14 @@
 class Solution {
     public int maxFrequency(int[] nums, int k) {
         Arrays.sort(nums);
-        int l = 0, res = 0;
-        long total = 0;
-
-        for (int r = 0; r < nums.length; r++) {
+        int l=0,res=0;
+        int total =0;
+        for(int r=0; r<nums.length; r++){
             total += nums[r];
-            
-            // While the cost to make all elements in window equal to nums[r] > k
-            // Cost = (Number of elements * Target value) - Current Sum
-            while ((long)nums[r] * (r - l + 1) - total > k) {
+            while((long) nums[r] * (r-l+1) - total > k){
                 total -= nums[l++];
             }
-            
-            res = Math.max(res, r - l + 1);
+            res = Math.max(res, r-l+1);
         }
         return res;
     }
